pyproject.toml
[project]
name = "excel-rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.7.4",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-text-splitters>=0.2.0",
    "openai>=1.30.0",
    "openpyxl>=3.1.2",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.35.0",
    "tabulate>=0.9.0",
]
